+++ asserts/js/docx/smoke/api_table_pr/table_layout/autofit.js
@@ -1,0 +1,17 @@
+builder.CreateFile("docx");
+var oDocument = Api.GetDocument();
+var oParagraph, oTable, oTableStyle, oTablePr;
+oParagraph = oDocument.GetElement(0);
+oParagraph.AddText("We set the table cells to preserve their size:");
+oTableStyle = oDocument.CreateStyle("CustomTableStyle", "table");
+oTableStyle.SetBasedOn(oDocument.GetStyle("Bordered - Accent 5"));
+oTablePr = oTableStyle.GetTablePr();
+oTable = Api.CreateTable(3, 3);
+oTablePr.SetTableLayout("autofit");
+oTable.SetTableLook(true, true, true, true, false, false);
+oTable.SetStyle(oTableStyle);
+var oCell = oTable.GetRow(0).GetCell(0);
+oCell.GetContent().GetElement(0).AddText("Fixed layout");
+oDocument.Push(oTable);
+builder.SaveFile("docx", "SetTableLayout.docx");
+builder.CloseFile();