+++ spec/xlsx/smoke/api_range/api_range_column_row_size_spec.rb
@@ -4,37 +4,31 @@ require 'spec_helper'
 
 describe 'ApiRange Set and get column/row size' do
   it 'ApiRange | SetColumnWidth method' do
-    skip if builder.semver < Semantic::Version.new('5.2.0')
     xlsx = builder.build_and_parse('asserts/js/xlsx/smoke/api_range/set_column_width.js')
     expect(xlsx.worksheets.first.columns.first.width.to_i).to eq(15)
   end
 
   it 'ApiRange | SetRowHeight method' do
-    skip if builder.semver < Semantic::Version.new('5.2.0')
     xlsx = builder.build_and_parse('asserts/js/xlsx/smoke/api_range/set_row_height.js')
     expect(xlsx.worksheets.first.rows.first.height).to eq(OoxmlParser::OoxmlSize.new(15, :point))
   end
 
   it 'ApiRange | GetRowHeight method' do
-    skip if builder.semver < Semantic::Version.new('5.2.0')
     xlsx = builder.build_and_parse('asserts/js/xlsx/smoke/api_range/get_row_height.js')
     expect(xlsx.worksheets.first.rows.first.cells.first.text).to eq('15')
   end
 
   it 'ApiRange | GetColumnWidth method' do
-    skip if builder.semver < Semantic::Version.new('5.2.0')
     xlsx = builder.build_and_parse('asserts/js/xlsx/smoke/api_range/get_column_width.js')
     expect(xlsx.worksheets.first.rows.first.cells.first.text).to eq('15')
   end
 
   it 'ApiRange | ColumnWidth getter' do
-    skip if builder.semver < Semantic::Version.new('5.2.0')
     xlsx = builder.build_and_parse('asserts/js/xlsx/smoke/api_range/column_width_getter.js')
     expect(xlsx.worksheets.first.rows.first.cells.first.text).to eq('15')
   end
 
   it 'ApiRange | RowHeight getter' do
-    skip if builder.semver < Semantic::Version.new('5.2.0')
     xlsx = builder.build_and_parse('asserts/js/xlsx/smoke/api_range/row_height_getter.js')
     expect(xlsx.worksheets.first.rows.first.cells.first.text).to eq('30')
   end